Fixes for building runtime on linux x86-64.
git-svn-id: https://llvm.org/svn/llvm-project/klee/trunk@77825 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/POSIX/fd_32.c
@@ -142,8 +142,8 @@ ssize_t getdents(int fd, struct dirent *dirp, size_t nbytes) {
     struct dirent64 *end = (struct dirent64*) ((char*) dp64 + res);
     while (dp64 < end) {
       struct dirent *dp = (struct dirent *) dp64;
-      unsigned name_len = (dp64->d_reclen - 
-                           (unsigned) &((struct dirent64*) 0)->d_name);
+      size_t name_len = (dp64->d_reclen - 
+                           (size_t) &((struct dirent64*) 0)->d_name);
       dp->d_ino = dp64->d_ino;
       dp->d_off = dp64->d_off;
       dp->d_reclen = dp64->d_reclen;
